@@ -1,6 +1,6 @@
 {
   "name": "crawler-client",
-  "version": "0.1.1",
+  "version": "0.2.0",
   "private": true,
   "scripts": {
     "serve": "vue-cli-service serve",
@@ -8,50 +8,37 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "axios": "^0.24.0",
+    "@element-plus/icons-vue": "^1.1.4",
+    "@vee-validate/i18n": "^4.5.10",
+    "@vee-validate/rules": "^4.5.10",
+    "axios": "^0.26.1",
     "core-js": "^3.18.1",
-    "element-ui": "^2.15.6",
+    "element-plus": "^2.1.8",
     "moment": "^2.29.1",
-    "vee-validate": "^3.4.13",
-    "vue": "^2.6.14",
-    "vue-i18n": "^8.26.5",
-    "vue-router": "^3.5.3",
-    "vuex": "^3.6.2",
-    "vuex-persistedstate": "^4.1.0"
+    "vee-validate": "^4.5.10",
+    "vue": "^3.2.31",
+    "vue-i18n": "^9.1.9",
+    "vue-router": "^4.0.14",
+    "vuex": "^4.0.2",
+    "vuex-persistedstate": "^4.1.0",
+    "webpack": "^5.71.0"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^4.5.13",
-    "@vue/cli-plugin-eslint": "^4.5.13",
-    "@vue/cli-service": "^4.5.13",
-    "babel-eslint": "^10.1.0",
-    "eslint-plugin-vue": "^8.1.1",
+    "@babel/core": "^7.12.16",
+    "@babel/eslint-parser": "^7.12.16",
+    "@vue/cli-plugin-babel": "^5.0.4",
+    "@vue/cli-plugin-eslint": "~5.0.0",
+    "@vue/cli-service": "^5.0.4",
+    "@vue/compiler-sfc": "^3.2.31",
+    "eslint": "^8.12.0",
+    "eslint-plugin-vue": "^8.0.3",
     "sass": "^1.42.1",
-    "sass-loader": "^10.2.0",
-    "vue-eslint-parser": "^8.0.1",
-    "vue-template-compiler": "^2.6.14"
-  },
-  "eslintConfig": {
-    "root": true,
-    "env": {
-      "node": true
-    },
-    "extends": [
-      "plugin:vue/essential",
-      "eslint:recommended"
-    ],
-    "rules": {},
-    "parserOptions": {
-      "parser": "babel-eslint"
-    }
-  },
-  "postcss": {
-    "plugins": {
-      "autoprefixer": {}
-    }
+    "sass-loader": "^12.6.0"
   },
   "browserslist": [
     "> 1%",
     "last 2 versions",
-    "not ie <= 8"
+    "not dead",
+    "not ie 11"
   ]
 }
