@@ -1,23 +1,18 @@
-import Vue from 'vue'
-import Vuex from 'vuex'
+import { createStore } from 'vuex'
 import CreatePersistedState from "vuex-persistedstate"
 import auth from './modules/auth'
 import user from './modules/user'
 import novelSearch from './modules/novel/search'
 import novelDetail from './modules/novel/detail'
 
-Vue.use(Vuex)
-
-const debug = process.env.NODE_ENV !== 'production'
-
-export default new Vuex.Store({
+export default createStore({
   modules: {
     auth,
     user,
     novelSearch,
     novelDetail
   },
-  strict: debug,
+  strict: process.env.NODE_ENV !== 'production',
   plugins: [CreatePersistedState({
     key: 'crawler-client',
     storage: window.sessionStorage
