@@ -1,11 +1,9 @@
 <template>
-  <div class="home">
-    <HelloWorld msg="Welcome to Crawler Client Vue.js App"/>
-  </div>
+  <img alt="Vue logo" src="@/assets/logo.png">
+  <HelloWorld msg="Welcome to Your Vue.js App"/>
 </template>
 
 <script>
-// @ is an alias to /src
 import HelloWorld from '@/components/HelloWorld.vue'
 
 export default {
