@@ -1,102 +1,95 @@
 <template>
-  <ValidationObserver
-    ref="observer"
-    v-slot="{ passes }"
+  <Form
+    as="el-form"
+    :validation-schema="schema"
+    @submit="onSubmit"
   >
-    <el-form
-      ref="form"
-      class="login"
+    <h2>Sign up</h2>
+    <Field
+      name="username"
+      v-slot="{ value, field, errorMessage }"
     >
-      <h2>Sign up</h2>
-      <ValidationProvider
-        name="username"
-        rules="required|max:16"
-        v-slot="{ errors }"
+      <el-form-item
+        :error="errorMessage"
+        class="input-form-wrapper"
       >
-        <el-form-item
-          :error="errors[0]"
-          class="input-form-wrapper"
-        >
-          <el-input
-            type="text"
-            placeholder="Username"
-            v-model="username"
-          />
-        </el-form-item>
-      </ValidationProvider>
-      <ValidationProvider
-        name="password"
-        rules="required|max:16"
-        v-slot="{ errors }"
+        <el-input
+          type="text"
+          placeholder="Username"
+          v-bind="field"
+          :model-value="value"
+        />
+      </el-form-item>
+    </Field>
+    <Field
+      name="password"
+      v-slot="{ value, field, errorMessage }"
+    >
+      <el-form-item
+        :error="errorMessage"
+        class="input-form-wrapper"
       >
-        <el-form-item
-          :error="errors[0]"
-          class="input-form-wrapper"
-        >
-          <el-input
-            type="password"
-            placeholder="Password"
-            v-model="password"
-          />
-        </el-form-item>
-      </ValidationProvider>
-      <ValidationProvider
-        name="email"
-        rules="required|max:64|email"
-        v-slot="{ errors }"
+        <el-input
+          type="password"
+          placeholder="Password"
+          v-bind="field"
+          :model-value="value"
+        />
+      </el-form-item>
+    </Field>
+    <Field
+      name="email"
+      v-slot="{ value, field, errorMessage }"
+    >
+      <el-form-item
+        :error="errorMessage"
+        class="input-form-wrapper"
       >
-        <el-form-item
-          :error="errors[0]"
-          class="input-form-wrapper"
-        >
-          <el-input
-            type="text"
-            placeholder="E-Mail"
-            v-model="email"
-          />
-        </el-form-item>
-      </ValidationProvider>
-      <el-button
-        type="primary"
-        @click="passes(signup)"
-      >Signup</el-button>
-      <p>Do you have an account?
-        <router-link to="/login">Login now!!</router-link>
-      </p>
-    </el-form>
-  </ValidationObserver>
+        <el-input
+          type="text"
+          placeholder="E-Mail"
+          v-bind="field"
+          :model-value="value"
+        />
+      </el-form-item>
+    </Field>
+    <el-button
+      type="primary"
+      native-type="submit"
+    >Signup</el-button>
+    <p>Do you have an account?
+      <router-link to="/login">Login now!!</router-link>
+    </p>
+  </Form>
 </template>
 
-<script>
+<script setup>
+import { useStore } from 'vuex'
+import { useRouter } from 'vue-router'
+import { Field, Form } from "vee-validate";
+import { ElMessage } from 'element-plus'
 import { USER_SIGNUP } from '@/store/actions/user'
-import { ValidationProvider, ValidationObserver } from "vee-validate"
 
-export default {
-  data: function () {
-    return {
-      username: '',
-      password: '',
-      email: ''
-    }
-  },
-  components: {
-    ValidationProvider,
-    ValidationObserver
-  },
-  methods: {
-    signup: function () {
-      const { username, password, email } = this
-      this.$store.dispatch(USER_SIGNUP, { username, password, email }).then(() => {
-        this.$router.push('/')
-      }).catch(error => {
-        this.$message({
-          showClose: true,
-          message: error,
-          type: 'error'
-        })
-      })
-    }
-  }
+const store = useStore()
+const router = useRouter()
+
+const schema = {
+  username: 'required|max:16',
+  password: 'required|max:16',
+  email: 'required|max:64|email'
+}
+
+function onSubmit(values) {
+  const { username, password, email } = values
+  store.dispatch(USER_SIGNUP, { username, password, email }).then(() => {
+    router.push('/')
+  }).catch(error => {
+    ElMessage({
+      message: error,
+      grouping: true,
+      type: 'error'
+    })
+  })
 }
 </script>
 
